requirements.txt
asgiref==3.3.1
Django==3.1.4
django-taggit==1.2.0
pkg-resources==0.0.0
pytz==2020.4
sqlparse==0.4.1
